+++ src/main/functions/ClientModuleOps.ps1
@@ -2,8 +2,7 @@
 
 $PATCH_FILE = ".\EntClient.dll"
 $PATCH_HASH = "1A1816C78925E734FCA16974BDBAA4AA"
-# $clientfilehost = "https://github.com/neuralpain/qbactivator/files/10475450/qb.patch"
-$clientfilehost = "https://raw.githubusercontent.com/neuralpain/qbactivator/v0.20.1/src/bin/ecc/EntClient.dll"
+$clientfilehost = "https://raw.githubusercontent.com/neuralpain/qbactivator/v0.20.2/src/bin/ecc/EntClient.dll"
 $CLIENT_MODULE = "$env:SystemRoot\Microsoft.NET\assembly\GAC_MSIL\Intuit.Spc.Map.EntitlementClient.Common\v4.0_8.0.0.0__5dc4fe72edbcacf5\Intuit.Spc.Map.EntitlementClient.Common.dll"
 
 function Get-ClientModule {
@@ -64,7 +63,7 @@ function Install-ClientModule {
   Write-Host "Done"
 }
 
-<### DATA MODULE ###>
+<#--- DATA MODULE ---#>
 
 $CLIENT_MODULE_DATA_PATH = "$env:ProgramData\Intuit\Entitlement Client\v8"
 $CLIENT_MODULE_DATA = "$CLIENT_MODULE_DATA_PATH\EntitlementDataStore.ecml"
@@ -86,13 +85,19 @@ function Remove-ClientDataModulePatch {
     Write-Host "No data patch found. Client data module was not modified."
     Start-Sleep -Milliseconds $TIME_NORMAL
   }
-
+  
   Write-NextOperationMenu
 }
 
+
 function Install-ClientDataModule {
   param ($Version)
-
+  
+  <# 
+    `Install-ClientDataModule` is still in development and is not 
+    functional enough to be included in an official release
+  #>
+  
   Remove-ClientDataModulePatch # if previously patched
   if (Test-Path $CLIENT_MODULE_DATA -PathType Leaf) { Copy-Item $CLIENT_MODULE_DATA "${CLIENT_MODULE_DATA}.bak" }