+++ src/main/functions/InformationDisplay.ps1
@@ -18,22 +18,35 @@ function Write-InfoLink {
   Write-Host "https://github.com/neuralpain/qbactivator" -ForegroundColor Green
 }      
 
+function Write-FileNotFound($File) {
+  Clear-Host
+  Write-Host "`nThe requested file could not be downloaded." -ForegroundColor White -BackgroundColor DarkRed
+  Write-Host "`nThe file was not found on the server at `"$File`"" -ForegroundColor White
+  Write-Host "Please submit this issue to @neuralpain. Thank you." -ForegroundColor White
+  Write-InfoLink
+}
+
 function Write-MainMenu {
   Clear-Host
   Write-HeaderLabel
   Write-Host "`nSelect QuickBooks product"
   Write-Host "-------------------------"
-  Write-Host "1 - Point of Sale (Single/Client)"
-  Write-Host "2 - Point of Sale (Server)"
-  Write-Host "3 - Other QuickBooks Activation"
+  Write-Host "1 - POS Store 1 (Server/Client)"
+  Write-Host "2 - POS Store 2 (Server/Client)"
+  Write-Host "3 - General QuickBooks activation"
+  Write-Host "    ^^^^ Pro/Enterprise/Other ^^^^"
   Write-Host "0 - Exit"
   $query = Read-Host "`n#"
   
   switch ($query) {
     0 { Write-ExitActivator; exit $NONE }
     1 { Invoke-QuickBooksInstaller }
-    2 { Invoke-QuickBooksInstaller -Server }
-    3 { Write-OptionUnavailable; Write-MainMenu }
+    2 { 
+      $script:SECOND_STORE = $true
+      Invoke-QuickBooksInstaller
+    }
+    # 3 { Write-OptionUnavailable; Write-MainMenu }
+    3 { Clear-Host; Write-Host; Invoke-Activation -GeneralActivation }
     default { Write-MainMenu }
   }
 }